Create package fragments and transfer fragment-files.
